Add pe-puppetserver and toml-rb before acceptance tests
Prior to this commit, the acceptance test setup was failing to apply the
influxdb::profile::toml class as it requires the pe-puppetserver
service. It also needed the toml-rb to be installed in the puppet gem
space. This commit adds the pe-puppetserver service resource to the
puppet apply and includes the toml-rb gem for puppet_gem.