Make Grafana datasource idempotent
Prior to this commit, using the grafana_datasource resource with "secure
json" parameters was not idempotent.  This is because the api obfuscates
these values, so we can't check the state of the datasource using this
resource.

This commit uses the "provisioning" directory to set the datasource at
service startup.  It uses a template to create the file so the token can
be a Sensitive type and be redacted from logs and reports.

It also adds an exec to make sure Grafana is ready to accept
connections.