(SUP-3201) Check port availability with systemd
Prior to this commit, we used an exec resource to check if the grafana
service were ready to receive connections.  However, this didn't account
for initial provisioning, and it's a bit of a hack.  This commit uses an
ExecStartPost directive instead, which is better because it allows us to
use metaparameters like 'require' directly on the service.