(SUP-3235) Use latest telegraf package on Ubuntu
Prior to this commit, Ubuntu provisioning would fail if the latest
telegraf package were newer than the $version parameter.  This looks to
be because only the latest package is available in the Ubuntu repo.
This commit changes the ensure property to be 'latest' on Ubuntu, and
changes the archive url to use the $version parameter.