(SUP-3396) Remove ha_last-sync-succeeded mbeans
Prior to this commit, including this mbean would cause an error when
trying to insert the Value field, because it is a boolean for this
mbean.  Since we don't need to plot this, this commit removes it.

It also adds 'error' and 'error_type' to excluded keys since we also
can't plot those.