(SUP-4065) Add parameter for group_by interval
This commit adds a dropdown for applying a group_by interval to all the
panels.  The default is 5 minutes to match the interval from the metrics
collector.  This makes it much easier to change the group_by, e.g. to 30
minutes to match the Puppet runinterval, or a custom collection
frequency.