Add state timelines for catalog and function perf
This commit adds a state timeline panel for operations performed as part
of catalog compilations.  It also replaces the "Function Performance"
panel with a state timeline.

Lastly, we removed the "Function Counts" panel.  The reason is that this
data is based on the "top 50" functions as determined by the aggregate metric.
This often creates large gaps in the graph and is largely useless.  See
https://tickets.puppetlabs.com/browse/TK-468