(SUP-4225) Deprecate Grafana archive installations
Prior to this commit, installing Grafana from an archive source was
listed as a supported method; however, this is currently non-functional
and none of the acceptance tests use this method.  Attempting to do so
will result in errors related to provisioned datasources, as well as not
starting grafana-server correctly.  This commit removes mentions of
Grafana archives from the README.